add insert and handleExceptions errror

+++ src/products/entities/product.entity.ts
@@ -12,7 +12,7 @@ export class Product {
     })
     title: string;
 
-    @Column('numeric', {
+    @Column('float', {
         default: 0
     })
     price: number;
@@ -28,7 +28,7 @@ export class Product {
     })
     slug: string;
 
-    @Column('int', {
+    @Column('float', {
         default: 0
     })
     stock: number;
